+++ tokenizers/src/models/bpe/serialization.rs
@@ -20,7 +20,7 @@ impl Serialize for BPE {
         model.serialize_field("end_of_word_suffix", &self.end_of_word_suffix)?;
 
         // Then the large ones
-        let mut merges: Vec<(&Pair, &u32)> = self
+        let mut merges: Vec<(&Pair, &u64)> = self
             .merges
             .iter()
             .map(|(pair, (rank, _))| (pair, rank))
@@ -72,7 +72,7 @@ impl<'de> Visitor<'de> for BPEVisitor {
         V: MapAccess<'de>,
     {
         let mut builder = BpeBuilder::new();
-        let mut vocab: Option<HashMap<String, u32>> = None;
+        let mut vocab: Option<HashMap<String, u64>> = None;
         let mut merges: Option<Vec<String>> = None;
         while let Some(key) = map.next_key::<String>()? {
             match key.as_ref() {
