+++ tokenizers/src/models/mod.rs
@@ -10,11 +10,11 @@ use std::collections::HashMap;
 /// Wraps a vocab mapping (ID -> token) to a struct that will be serialized in order
 /// of token ID, smallest to largest.
 struct OrderedVocabIter<'a> {
-    vocab_r: &'a HashMap<u32, String>,
+    vocab_r: &'a HashMap<u64, String>,
 }
 
 impl<'a> OrderedVocabIter<'a> {
-    fn new(vocab_r: &'a HashMap<u32, String>) -> Self {
+    fn new(vocab_r: &'a HashMap<u64, String>) -> Self {
         Self { vocab_r }
     }
 }
@@ -24,7 +24,7 @@ impl<'a> Serialize for OrderedVocabIter<'a> {
     where
         S: Serializer,
     {
-        let iter = (0u32..(self.vocab_r.len() as u32)).map(|i| (&self.vocab_r[&i], i));
+        let iter = (0u64..(self.vocab_r.len() as u64)).map(|i| (&self.vocab_r[&i], i));
         serializer.collect_map(iter)
     }
 }
