+++ tokenizers/src/models/wordlevel/mod.rs
@@ -29,7 +29,7 @@ impl fmt::Display for Error {
 }
 
 struct Config {
-    vocab: HashMap<String, u32>,
+    vocab: HashMap<String, u64>,
     unk_token: String,
 }
 
@@ -57,7 +57,7 @@ impl WordLevelBuilder {
     }
 
     /// Set the vocab (token -> ID) mapping.
-    pub fn vocab(mut self, vocab: HashMap<String, u32>) -> Self {
+    pub fn vocab(mut self, vocab: HashMap<String, u64>) -> Self {
         self.config.vocab = vocab;
         self
     }
@@ -86,8 +86,8 @@ impl WordLevelBuilder {
 
 #[derive(PartialEq)]
 pub struct WordLevel {
-    vocab: HashMap<String, u32>,
-    vocab_r: HashMap<u32, String>,
+    vocab: HashMap<String, u64>,
+    vocab_r: HashMap<u64, String>,
     unk_token: String,
 }
 
@@ -120,7 +120,7 @@ impl WordLevel {
             Value::Object(m) => {
                 for (token, id) in m {
                     if let Value::Number(id) = id {
-                        let id = id.as_u64().ok_or(Error::BadVocabulary)? as u32;
+                        let id = id.as_u64().ok_or(Error::BadVocabulary)? as u64;
                         vocab.insert(token, id);
                     }
                 }
@@ -156,7 +156,7 @@ impl Model for WordLevel {
                     .ok_or(Error::MissingUnkToken)?,
                 value: token,
                 offsets: initial_offsets,
-                word: index as u32,
+                word: index as u64,
             };
 
             output_tokens.push(t);
@@ -165,15 +165,15 @@ impl Model for WordLevel {
         Ok(output_tokens)
     }
 
-    fn token_to_id(&self, token: &str) -> Option<u32> {
+    fn token_to_id(&self, token: &str) -> Option<u64> {
         self.vocab.get(token).copied()
     }
 
-    fn id_to_token(&self, id: u32) -> Option<&str> {
+    fn id_to_token(&self, id: u64) -> Option<&str> {
         self.vocab_r.get(&id).map(String::as_ref)
     }
 
-    fn get_vocab(&self) -> &HashMap<String, u32> {
+    fn get_vocab(&self) -> &HashMap<String, u64> {
         &self.vocab
     }
